split range logic into several files
